// backend/yamailer/src/app.schema.ts
export type AppConfigSchema = {
    PORT: number;
    MONGO_CONNECT: string;
    MONGO_NAME: string;
    CLIENT_ID: string;
    CLIENT_SECRET: string;
    REDIRECT_URI: string;
    SERVER_URI: string;
    GOOGLE_CLIENT_ID: string;
    GOOGLE_CLIENT_SECRET: string;
    SESSIONS_SECRET_KEY: string;
    YANDEX_CLIENT_ID: string;
    YANDEX_CLIENT_SECRET: string;
    MAILRU_CLIENT_ID: string;
    MAILRU_CLIENT_SECRET: string;
};